requirements.txt
flask

gunicorn

line-bot-sdk==1.3.0
imgurpython==1.1.7
bs4